% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_normals.R
\name{add_normals}
\alias{add_normals}
\title{Write Normals of Scene Object}
\usage{
add_normals(x)
}
\arguments{
\item{x}{scene object (object of class "scenesetr_obj")}
}
\value{
Updated scene object.
}
\description{
Add or edit the surface normal vectors of a scene object.
}
\details{
Uses Newell's method to calculate the normal vector of each polygon.
The updated normal vectors are stored in \code{x$normals}.
The index of a normal vector is added to each polygon in the first row of \code{x$faces}.

Used in \link{read_obj} if normals cannot be found, and in \link{st_as_obj}. Useful if
normals have been misread by \link{read_obj}, for example in the case that vertex
normals rather than surface normals are supplied in the .obj file. Also useful
if points or faces have been modified after scene object creation.
}
