% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.R
\name{aspect}
\alias{aspect}
\alias{aspect<-}
\title{The Aspect Ratio of a Camera}
\usage{
aspect(x)

aspect(x) <- value
}
\arguments{
\item{x}{camera (object of class "scenesetr_camera")}

\item{value}{numeric replacement value}
}
\value{
For \code{aspect(x)}, the aspect ratio as a double.

For \code{aspect(x) <- value}, the updated camera. (Note that the value of
\code{aspect(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get or set the aspect ratio of a camera.
}
\details{
\code{aspect(x)} returns the aspect ratio of a camera.

\code{aspect(x) <- value} assigns a value to the aspect ratio of a camera.

By default, \code{\link[=camera]{camera()}} generates a camera with an aspect ratio of 1.

The aspect ratio of a camera in a scene is used by \code{\link[=see]{see()}} in setting
the height of the \code{\link[grDevices:x11]{grDevices::X11()}} graphics device to seven inches divided by
the aspect ratio.

Using a behaviour to change the aspect ratio of a camera will not change the
height of the graphics device until \code{\link[=see]{see()}} is called again on the scene.
}
