% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaves.R
\name{behaves}
\alias{behaves}
\alias{\%behaves\%}
\title{Query Behaviors of Scene Elements}
\usage{
behaves(x, behavior)

x \%behaves\% behavior
}
\arguments{
\item{x}{scene element or scene. Camera (object of class "scenesetr_camera")
or light (object of class "scenesetr_light") or scene object (object of class
"scenesetr_obj") or scene (object of class "scenesetr_scene").}

\item{behavior}{character variable or variable name}
}
\value{
Boolean indicating whether the scene element has the behavior.
}
\description{
Tests if a named behavior is present in the behaviors of a scene element.
If a scene is supplied, \code{behaves()} will return a logical vector,
one for each element in the scene. \verb{\%behaves\%} and \code{behaves()} are synonyms.
}
\details{
The \code{behavior} argument can take either a character variable of the name of
the behavior to be queried, or a variable name to be read by \code{\link[=match.call]{match.call()}}.
This is checked for equality against the names of behavior list of the
scene element.

The logical vector output of \code{behaves()} when called on a scene can be used
to subset a scene.
}
\seealso{
\code{\link[=behave]{behave()}}, \code{\link[=spin]{spin()}}.
}
