% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviours.R
\name{behaviours}
\alias{behaviours}
\alias{behaviours<-}
\alias{behave}
\title{The Behaviours of a Scene Element}
\usage{
behaviours(x)

behaviours(x) <- value

behave(x, ...)
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{named list of replacement behaviour functions}

\item{...}{behaviour functions}
}
\value{
For \code{behaviours(x)}, a named list of behaviour functions.

For \code{behaviours(x) <- value}, the updated element. (Note that the value of
\code{behaviours(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)

For \code{behaves(x, ...)}, an element with modified behaviours.
}
\description{
Functions to get or set the behaviours of a scene element.
}
\details{
The behaviours of an element are a named list of functions describing how
an element is modified every frame when \code{\link[=record]{record()}} is called on a scene
containing the element. Behaviours are applied to an element each frame
in the order in which they appear in the named list.

\code{behaviours(x)} returns the behaviours of a scene element.

\code{behaviours(x) <- value} assigns a value to the behaviours of a scene element.
Only recommended in setting the behaviours of an element to (a subset of)
the behaviours of another element.

\code{behave(x, ...)} returns a scene element with updated behaviours.
The functions provided in \code{...} are added to the named list of behaviours
of the scene element. The name of a new behaviour is set to the name
of the function as captured  by \code{\link[=match.call]{match.call()}}. It is therefore recommended
to specify each behaviour using the name of a predefined variable.

A behaviour can be defined by a user, with care to adhere to the following
principles:
\itemize{
\item The arguments of the function must include \code{...} in addition to the following
optional parameters. \code{element} is the scene element immediately before the
behaviour is applied. \code{scene} is the entire scene as it was the previous frame.
\code{keys} is a character vector of the keys held during
the current frame. \code{last_keys} is a character vector of the keys held during
the previous frame. \code{frame} is the frame number of the current frame.
\item The function must return the modified scene element. The returned element
is then rendered if the behaviour is the last in the named list, or is passed
to the next behaviour in the list as the \code{element} argument.
}
}
\seealso{
\code{\link[=behaves]{behaves()}}, \code{\link[=spin]{spin()}}.
}
