% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\alias{border<-}
\alias{set_border}
\title{The Borders of Scene Objects}
\usage{
border(x)

border(x) <- value

set_border(x, value)
}
\arguments{
\item{x}{scene object (object of class "scenesetr_camera")}

\item{value}{replacement value}
}
\value{
For \code{border(x)}, the border value of the scene object

For \code{border(x) <- value}, and \code{set_border(x, value)}, the updated scene object.
(Note that the value of
\code{border(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get or set the border properties of scene objects.
}
\details{
\code{border(x)} returns the border value of the scene object.

\code{border(x) <- value} assigns a border value to a scene object.

\code{set_border} is an alias of \verb{border<-} useful in pipes.

By default, \code{\link[=read_obj]{read_obj()}} and \code{\link[=st_as_obj]{st_as_obj()}} generate scene objects with a
border value of \code{NA}, and \code{\link[=point_cloud]{point_cloud()}} a border value of \code{"white"}.

Using \code{\link[=paint]{paint()}} on a point cloud will not affect its appearance. To change
the colours of a point cloud, the border value must be set, as each point
is rendered as a polygon.

The border value of a scene element is passed to \code{\link[graphics:polygon]{graphics::polygon()}} in
\code{\link[=record]{record()}}. \code{value} can therefore either be length one or a vector with an
element for each polygon in the scene object.
}
