% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube_obj}
\alias{cube_obj}
\alias{half_cube_obj}
\alias{pyramid_obj}
\title{Geometric Scene Objects}
\usage{
cube_obj(side_length = 1, inverse = FALSE)

half_cube_obj(side_length = 1, inverse = FALSE)

pyramid_obj(side_length = 1, inverse = FALSE)
}
\arguments{
\item{side_length}{side length of each square face}

\item{inverse}{logical; should normals face inwards?}
}
\value{
Scene object (object of class "scenesetr_obj")
painted white with no position or behavior.
}
\description{
Create a cubic or pyramidal scene object.
}
\details{
Create a cube within the interval \verb{[0, side_length]} in each direction,
or a square-based pyramid \verb{[-side_length, side_length]} in each direction
with an apex at \verb{(0, 0, side_length)} if placed at the origin.

\code{half_cube_obj()} returns the three faces of a \code{cube_obj()}
that meet at the origin.

\code{inverse} may be a logical vector recycled over each normal.
}
\seealso{
\code{\link[=read_obj]{read_obj()}}
}
