% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\alias{direction<-}
\title{The Direction of a Scene Element}
\usage{
direction(x)

direction(x) <- value
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{numeric replacement vector. 3-D (x,y,z) coordinates}
}
\value{
For \code{direction(x)}, the direction vector of the scene element.

For \code{direction(x) <- value}, the updated scene element. (Note that the value of
\code{direction(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get and set the direction a scene element faces.
}
\details{
The rotation of a scene element is stored as a quaternion.

\code{direction(x)} returns a direction vector of unit length in 3-D (x,y,z)
coordinates derived from the rotation quaternion of an element.

\code{direction(x) <- value} assigns the result of \code{\link{point}}\verb{(x, value)}.
}
\seealso{
\code{\link[=point]{point()}}, \code{\link[=rotate]{rotate()}}, \code{\link[=rotation]{rotation()}}, \code{\link[=looked_at]{looked_at()}}, \code{\link[=skewer]{skewer()}}.
}
