% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenland_bed.R
\docType{data}
\name{greenland_bed}
\alias{greenland_bed}
\alias{greenland_ice}
\title{Greenland Relief Data}
\format{
An object of class \code{stars} with 1886 rows and 858 columns.

An object of class \code{stars} with 1200 rows and 400 columns.
}
\source{
\url{https://data.noaa.gov/onestop/collections/details/986f93b8-0a07-4e44-a2bc-bdee3c02fd5a}

\code{\link[stars:st_downsample]{stars::st_downsample()}} was performed with \code{n = 20} and relief in meters was
converted to kilometers.

NOAA National Centers for Environmental Information.
2022: ETOPO 2022 15 Arc-Second Global Relief Model.
NOAA National Centers for Environmental Information.
DOI: 10.25921/fd45-gt74. Accessed 01/03/2024.
}
\usage{
greenland_bed

greenland_ice
}
\description{
A subset of data from the ETOPO 2022 15 Arc-Second Global Relief Model.
\code{greenland_bed} is bed relief and \code{greenland_ice} is ice surface relief.
}
\keyword{datasets}
