% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_range.R
\name{in_range}
\alias{in_range}
\title{Test the Proximity of Scene Elements}
\usage{
in_range(x, y, range)
}
\arguments{
\item{x, y}{scene elements. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{range}{numeric distance value}
}
\value{
Logical value indicating whether the elements are in range.
}
\description{
Test whether two scene elements are within a specified distance in 3-D space.
}
\details{
Tests whether the distance between the positions of two elements is less
than a threshold distance specified by \code{range}.
The position of an element can be queried with \code{\link[=position]{position()}}.

Useful in defining collision behaviours.

If an element is unplaced, \code{in_range()} will return \code{FALSE}.
}
\seealso{
\code{\link[=looked_at]{looked_at()}}, \code{\link[=behave]{behave()}}, \code{\link[=spin]{spin()}}.
}
