% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{light}
\alias{light}
\title{Create a Light}
\usage{
light(position = NA, direction = NA, color = "white")
}
\arguments{
\item{position}{numeric vector. 3-D (x,y,z) coordinates}

\item{direction}{numeric vector. 3-D (x,y,z) coordinates}

\item{color}{passed to \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}}
}
\value{
Object of class "scenesetr_light".
}
\description{
Create an object of class "scenesetr_light". A light illuminates a scene to
be seen from the viewpoint of a camera when \code{\link[=record]{record()}} is called on the scene.
}
\details{
There are four categories of light. Ambient light, directional light, point
light and spotlight. Default arguments return a white ambient light with no behaviour.

If a light is unplaced and has no direction, it is an ambient light, which
exerts equal shading on every polygon.

If a light is unplaced and has a direction, it is a directional light, which
acts as a uniform parallel light over the whole scene, similar to the sun.

If a light has a location and no direction, it is a point light, which acts
as an omni-directional light from a point source at its location.

If a light has both a location and a direction, it is a spotlight, which acts
as a point light, but only acting on a cone in the direction specified.

\code{direction} will be normalised to a 3-D vector of unit length.
}
\seealso{
\code{\link[=camera]{camera()}}, \code{\link[=scene]{scene()}}.
}
