% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{location}
\alias{location}
\alias{location<-}
\title{The Location of a Scene Element}
\usage{
location(x)

location(x) <- value
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{numeric replacement vector. 3-D (x,y,z) coordinates}
}
\value{
For \code{location(x)}, the location vector of the scene element.

For \code{location(x) <- value}, the updated scene element. (Note that the value of
\code{location(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get and set the location of a scene element.
}
\details{
\code{location(x)} returns a location vector in 3-D (x,y,z) coordinates.

\code{location(x) <- value} assigns the result of \code{\link{place}}\verb{(x, value)}.
}
