% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looked_at.R
\name{looked_at}
\alias{looked_at}
\title{Observation of a Scene Element}
\usage{
looked_at(lookee, looker, angle = 15)
}
\arguments{
\item{lookee, looker}{scene elements. Camera (object of class
"scenesetr_camera") or light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{angle}{numeric value in degrees}
}
\value{
Logical value indicating whether the \code{lookee} is looked at by the \code{looker}.
}
\description{
Test whether a scene element is facing another scene element with a threshold
angle of observation.
}
\details{
Tests whether the location of the \code{lookee} is within a cone from the location
of the \code{looker} in the direction of the \code{looker} with a specified threshold
\code{angle} from the line of sight of the \code{looker}.

The location and direction of an element can be queried with \code{\link[=location]{location()}} and
\code{\link[=direction]{direction()}} respectively.

Useful in defining behaviours on sight or interaction.

If an element is unplaced, \code{looked_at()} will return \code{FALSE}.
}
