% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_example.R
\name{obj_example}
\alias{obj_example}
\title{Example Object Files}
\usage{
obj_example(path = NULL)
}
\arguments{
\item{path}{character vector specifying file(s)}
}
\value{
Character vector of all file names in the "extdata" subdirectory of
the "scenesetr" package, or a character vector of full file names of queried
example object files in the "extdata" subdirectory.
}
\description{
Find the full file name of an example object file or produce a character
vector of the names of every example object file in the "scenesetr" package.
}
\details{
If no \code{path} is provided, \code{obj_example()} returns a character
vector of the names of every example object file in the "scenesetr" package.
If a \code{path} character vector matching one or more elements of the returned
vector of \code{obj_example()} is provided, \code{obj_example(path)} returns the full
file name(s) of the example object file(s) provided.

Calls \code{\link[base:system.file]{base::system.file()}} with the "extdata" subdirectory and "scenesetr"
package.
}
