% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{orientation}
\alias{orientation}
\alias{orientation<-}
\title{The orientation of a Scene Element}
\usage{
orientation(x)

orientation(x) <- value
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{numeric replacement vector.
Quaternion or an R object containing \code{NA}.}
}
\value{
For \code{orientation(x)}, the quaternion of the scene element as a numeric vector.

For \code{orientation(x) <- value}, the updated scene element. (Note that the value of
\code{orientation(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get and set the orientation of a scene element.
}
\details{
The orientation of a scene element is stored as a quaternion describing the
rotation required to orient the object from its original orientation.
The original orientation of a scene element is facing the positive z direction.

\code{orientation(x)} returns the rotation quaternion of an element.

\code{orientation(x) <- value} assigns a rotation quaternion to an element.

Useful in copying orientation between elements:
\code{orientation(element2) <- orientation(element1)}. This is safer than
\code{direction(element2) <- direction(element1)} as it conserves roll.

If \code{anyNA(value)}, the updated scene element has no orientation.
}
\seealso{
\code{\link[=point]{point()}}, \code{\link[=rotate]{rotate()}}, \code{\link[=direction]{direction()}}, \code{\link[=looked_at]{looked_at()}}, \code{\link[=skewer]{skewer()}}.
}
