% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{The Colour of a Scene Element}
\usage{
paint(x, col)
}
\arguments{
\item{x}{scene object (object of class "scenesetr_obj") or light (object of
class "scenesetr_light") or scene (object of class "scenesetr_scene")}

\item{col}{vector of any of the three kinds of \R color specifications,
    i.e., either a color name (as listed by \code{\link[grDevices]{colors}()}), a
    hexadecimal string of the form \code{"#rrggbb"} or
    \code{"#rrggbbaa"} (see \code{\link[grDevices]{rgb}}), or a positive integer
    \code{i} meaning \code{\link[grDevices]{palette}()[i]}.}
}
\value{
Scene object or light or scene with updated colour.
}
\description{
Change the colour of a light or scene object.
}
\details{
Paints all polygons of a scene object a specified colour, or alternatively
sets the colour of a light.

Additionally, two colours can be given when painting a scene object, upon
which a gradient between the two colours is applied to the object based on
the degree to which each polygon faces the horizontal axis orthogonal to
the direction the scene object faces. This gradient is a linear interpolation
in the RGB colourspace.

A scene object must be painted to be rendered by \code{\link[=see]{see()}}.

By default, \code{\link[=light]{light()}} creates a white light.

If \code{x} is a scene, \code{paint()} will be applied to every element in the scene.
}
