% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place.R
\name{place}
\alias{place}
\title{Place a Scene Element}
\usage{
place(x, where)
}
\arguments{
\item{x}{scene element or scene. Camera (object of class "scenesetr_camera")
or light (object of class "scenesetr_light") or scene object (object of class
"scenesetr_obj") or scene (object of class "scenesetr_scene").}

\item{where}{numeric vector or an R object containing \code{NA}.
3-D (x,y,z) coordinates.}
}
\value{
Scene element or scene with updated location.
}
\description{
Set the location of a scene element in 3-D space.
}
\details{
If \code{anyNA(where)}, the scene element will be unplaced. Otherwise,
\code{where} must be a numeric vector of length three specifying the
new location of the element in 3-D coordinates (x,y,z).

An unplaced element will not be rendered by \code{\link[=record]{record()}}.

If \code{x} is a scene, \code{place()} will be applied to every element in the scene.
}
\seealso{
\code{\link[=move]{move()}}, \code{\link[=location]{location()}}.
}
