% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Point a Scene Element in a Direction}
\usage{
point(x, direction, rotate_to = FALSE)
}
\arguments{
\item{x}{scene element or scene. Camera (object of class "scenesetr_camera")
or light (object of class "scenesetr_light") or scene object (object of class
"scenesetr_obj") or scene (object of class "scenesetr_scene").}

\item{direction}{numeric vector. 3-D (x,y,z) coordinates}

\item{rotate_to}{logical value indicating whether the element should be
rotated to its new direction from its current orientation.}
}
\value{
Scene element or scene with updated rotation.
}
\description{
Set the direction of a scene element, either ensuring the element remains
upright, or, if \code{rotate_to} is \code{TRUE}, ensuring it takes one rotation about
an axis to transition from its current orientation to the specified direction.
}
\details{
The rotation of a scene element is stored as a quaternion.

The 3-D \code{direction} vector is normalised to unit length, and a quaternion is
assigned to the scene element describing its new rotation to face the direction.

By default, \code{point()} assigns a rotation quaternion with no roll, ensuring the
element is upright once rotated to face the direction. Alternatively, if
\code{rotate_to} is \code{TRUE}, the element is rotated about a single axis from its
current orientation to face the new direction, so may not remain upright.

An upright element is defined as having a quaternion that describes a
rotation about the vertical axis, followed by a rotation upwards or downwards
relative to the new direction of the element, so that no roll is described.

If \code{anyNA(direction)}, the returned element has no rotation. If the \code{direction}
given is \code{c(0,0,0)}, the element is set to face the positive z direction.

If \code{x} is a scene, \code{point()} will be applied to every element in the scene.
}
\seealso{
\code{\link[=direction]{direction()}}, \code{\link[=rotate]{rotate()}}, \code{\link[=rotation]{rotation()}}, \code{\link[=looked_at]{looked_at()}}, \code{\link[=skewer]{skewer()}}.
}
