% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_cloud.R
\name{point_cloud}
\alias{point_cloud}
\title{Point Cloud}
\usage{
point_cloud(x)
}
\arguments{
\item{x}{matrix or data frame of coordinates}
}
\value{
Scene object (object of class "scenesetr_obj").
}
\description{
The first three columns of \code{x} are taken to be the x, y and z coordinates
of each point in the cloud.

Returns a scene object of polygons, each with two identically located vertices,
so that each polygon is rendered as a point. As each polygon has zero area,
they are painted transparent and have their borders set to white.

To change the colour of the points, use \code{\link[=set_border]{set_border()}}. To change the size of
the points, use the \code{lwd} argument of \code{\link[=record]{record()}}.
}
\details{
Create a scene object comprising a cloud of points in 3-D space.
}
\seealso{
\code{\link[=st_as_obj]{st_as_obj()}}, \code{\link[=scene]{scene()}}, \code{\link[=record]{record()}}.
}
