% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position}
\alias{position}
\alias{position<-}
\title{The Position of a Scene Element}
\usage{
position(x)

position(x) <- value
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{numeric replacement vector. 3-D (x,y,z) coordinates}
}
\value{
For \code{position(x)}, the position vector of the scene element.

For \code{position(x) <- value}, the updated scene element. (Note that the value of
\code{position(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get and set the position of a scene element.
}
\details{
\code{position(x)} returns a position vector in 3-D (x,y,z) coordinates.

\code{position(x) <- value} assigns the result of \code{\link{place}}\verb{(x, value)}.
}
\seealso{
\code{\link[=place]{place()}}, \code{\link[=move]{move()}}, \code{\link[=looked_at]{looked_at()}}, \code{\link[=in_range]{in_range()}}.
}
