% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_device.R
\name{quit_device}
\alias{quit_device}
\alias{restart}
\title{Quit or Restart a Scene Render}
\usage{
quit_device(...)

restart(...)
}
\arguments{
\item{...}{passed to \code{\link[=cat]{cat()}}}
}
\value{
For \code{quit_device()}, 0.

For \code{restart()}, 1.
}
\description{
Functions to be called within behaviours to quit or restart \code{\link[=record]{record()}}.
}
\details{
A behaviour that returns the result of \code{quit_device()} or \code{restart()} instead
of returning the updated element causes \code{\link[=record]{record()}} to quit the graphics device
or return the scene to its original value respectively when it is applied.

Restarting a scene resets all values stored in its elements. See \code{\link{initial<-}}.
}
\examples{
count_to_thirty <- function(element, frame, ...){
  print(frame)
  if(frame == 30) return(quit_device("Thirty reached\n"))
  element
}

cam <- behave(camera(), count_to_thirty)
scene <- scene(cam)
\dontrun{
see(scene)
}

restart_if_twenty <- function(element, frame, ...){
  if(frame == 20) return(restart("Twenty reached\n"))
  element
}

restarting_cam <- behave(cam, restart_if_twenty)
restarting_scene <- scene(restarting_cam)
\dontrun{
# Restarting doesn't restart the frame number; only restarts the scene
see(restarting_scene)
}
}
\seealso{
\code{\link[=behave]{behave()}}, \code{\link[=spin]{spin()}}.
}
