% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obj.R
\name{read_obj}
\alias{read_obj}
\title{Data Input from Object File}
\usage{
read_obj(file, take_first = TRUE)
}
\arguments{
\item{file}{a connection object or a character string.}

\item{take_first}{logical value indicating whether only the first object
should read and returned}
}
\value{
Scene object (object of class "scenesetr_obj") if \code{take_first} is
\code{TRUE}, otherwise scene (object of class "scenesetr_scene") containing all
scene objects described in the file.
}
\description{
Create a scene object or a scene from a .obj file.
}
\details{
Object files can contain multiple objects. \code{take_first = FALSE} allows all
objects in an object file to be returned as a scene of named scene objects.
Otherwise, a single scene object is returned.

The vertices, face normals, face normal indices and face vertex indices of
an object are read. All face indices are stored in the same matrix, with
normal indices of each face in the first row.

If no face normals or face normal indices are found, they will be added
by \link{add_normals}.

All returned scene objects are painted white, unplaced, have no behaviours and
face the positive z direction.
}
\seealso{
\code{\link[=st_as_obj]{st_as_obj()}}, \code{\link[=scene]{scene()}}, \code{\link[=see]{see()}}.
}
