% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Interactive 3-D Render of a Scene}
\usage{
record(x, device = grDevices::dev.new, width = 7, render_order = NULL, ...)
}
\arguments{
\item{x}{scene (object of class "scenesetr_scene")
or recording (object of class "scenesetr_recording")}

\item{device}{function. Graphics device to be used.}

\item{width}{passed to device function}

\item{render_order}{optional numeric vector of object indices}

\item{...}{passed to device function}
}
\value{
Object of class "scenesetr_recording", invisibly. List of three elements:
\itemize{
\item \code{initial_scene}: the original scene passed to \code{record()},
\item \code{final_scene}: the scene as it was in the last frame before quitting the device,
\item \code{key_inputs}: a list of key inputs, with one element per frame recorded.
}
}
\description{
View a scene from the perspective of a camera. Record and replay how behaviours
play out frame by frame with real-time key inputs from the user.
}
\details{
Only available on Windows due to the method of obtaining key inputs. This
may change in the future.

Supplying a scene with no camera is an error. A scene with no lights
or no scene objects will appear blank.

Although key inputs can be used in the behaviours of scene elements, "R" and
"Q" are reserved for resetting the scene and quitting the device respectively.

\code{x} can be a recording returned from \code{record()} or \code{\link[=record_gif]{record_gif()}},
in which case, the initial scene and key inputs will be taken from the recording,
rather than using real-time key inputs.

If \verb{device = }\code{\link[grDevices:dev]{grDevices::dev.new}}, as default, \code{noRStudioGD = TRUE} is set,
as the RStudio graphics device does not accept arguments such as \code{width} and \code{height}.
The \code{height} argument passed to the device depends on \code{width} and the aspect of the
camera.

If \code{render_order} is specified, rendering order is no longer calculated based on
position in the scene. \code{render_order = 3:1}, for example, ensures only the first
three objects in the scene are rendered each frame, from last to first.
}
\seealso{
\code{\link[=scene]{scene()}}, \code{\link[=read_obj]{read_obj()}}, \code{\link[=record_gif]{record_gif()}}.
}
