% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{replay}
\alias{replay}
\title{Replay a Recording of a Scene}
\usage{
replay(recording, device = grDevices::X11, width = 7, ...)
}
\arguments{
\item{recording}{output of \code{\link[=see]{see()}}. Object of class "scenesetr_recording".}

\item{device}{function. Graphics device to be used.}

\item{width}{passed to device function}

\item{...}{passed to device function}
}
\value{
\code{recording}, unchanged.
}
\description{
Replay a recording from \code{\link[=see]{see()}} and re-record it on a new device.
}
\details{
Recomputes every frame of a recording using the same scene and key inputs
as were used in the recording. These frames are plotted in the specified
device. The \code{final_scene} element of the recording is not used, and is
provided in case it is of use to the user.

Useful for record gifs, by saving each frame as a png. For example,
\code{replay(recording, grDevices::png, 800, filename = "frames/frame_\%05d.png")}.
This can then be recorded as a gif using for example \code{\link[gifski:gifski]{gifski::gifski()}}.
}
