% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a Scene Element}
\usage{
rotate(x, axis, angle)
}
\arguments{
\item{x}{scene element or scene. Camera (object of class "scenesetr_camera")
or light (object of class "scenesetr_light") or scene object (object of class
"scenesetr_obj") or scene (object of class "scenesetr_scene").}

\item{axis}{numeric vector or character string.
3-D (x,y,z) coordinates or one of \code{c(up, down, left, right, clockwise)}.}

\item{angle}{numeric value in degrees}
}
\value{
Scene element or scene with updated rotation.
}
\description{
Rotate a scene element from its current orientation.
}
\details{
If \code{axis} is specified as a 3-D vector, the element is rotated about its
location about an axis parallel to the specified vector.

\code{axis} can be specified as a character string describing the direction in
which the object is to be rotated relative to its current direction.
See \code{\link[=skewer]{skewer()}} with \code{to_rotate = TRUE}.

If \code{x} is a scene, \code{rotate()} will be applied to every element in the scene.
}
\seealso{
\code{\link[=point]{point()}}, \code{\link[=direction]{direction()}}, \code{\link[=rotation]{rotation()}}, \code{\link[=looked_at]{looked_at()}}.
}
