% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\alias{rotation<-}
\title{The Rotation of a Scene Element}
\usage{
rotation(x)

rotation(x) <- value
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{value}{numeric replacement vector.
Quaternion or an R object containing \code{NA}.}
}
\value{
For \code{rotation(x)}, the quaternion of the scene element as a numeric vector.

For \code{rotation(x) <- value}, the updated scene element. (Note that the value of
\code{rotation(x) <- value} is that of the assignment, \code{value}, not the return
value from the left-hand side.)
}
\description{
Functions to get and set the rotation of a scene element.
}
\details{
The rotation of a scene element is stored as a quaternion.

\code{rotation(x)} returns the rotation quaternion of an element.

\code{rotation(x) <- value} assigns a rotation quaternion to an element.

Useful in copying rotation between elements:
\code{rotation(element2) <- rotation(element1)}. This is safer than
\code{direction(element2) <- direction(element1)} as it conserves roll.

If \code{anyNA(value)}, the updated scene element has no rotation.
}
\seealso{
\code{\link[=point]{point()}}, \code{\link[=rotate]{rotate()}}, \code{\link[=direction]{direction()}}, \code{\link[=looked_at]{looked_at()}}, \code{\link[=skewer]{skewer()}}.
}
