% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{scene}
\alias{scene}
\title{Create a Scene}
\usage{
scene(..., list = NULL)
}
\arguments{
\item{...}{scene elements. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj"). Passed to \code{\link[=list]{list()}}.}

\item{list}{optional list of scene elements}
}
\value{
Object of class ("scenesetr_scene").
}
\description{
Create an object of class "scenesetr_scene" containing scene elements.
}
\details{
A scene is a list of scene elements, so can be subset with
\code{[}, \code{[[} and \code{$} and combined with \code{c()}. A scene may have names.

A scene with lights, scene objects and a camera can be visualised with \code{\link[=record]{record()}}.
}
\seealso{
\code{\link[=camera]{camera()}}, \code{\link[=light]{light()}}, \code{\link[=record]{record()}}.
}
