% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Interactive 3-D Render of a Scene}
\usage{
see(scene)
}
\arguments{
\item{scene}{scene (object of class "scenesetr_scene")}
}
\value{
The scene as it was in the last frame before quitting the device,
invisibly.
}
\description{
View a scene from the perspective of a camera. Watch behaviours play out
frame by frame with real-time key inputs from the user.
}
\details{
Only available on windows due to the method of obtaining key inputs. This
may change in the future. The use of the \code{\link[grDevices:x11]{grDevices::X11()}} device imposes
additional platform restrictions.

A scene must have a camera for \code{see()} to function. A scene with no lights
or no scene objects will appear blank.

Although key inputs can be used in the behaviours of scene elements, "R" and
"Q" are reserved for resetting the scene and quitting the device respectively.
}
