% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewer.R
\name{skewer}
\alias{skewer}
\title{Which Way is Up?}
\usage{
skewer(
  x,
  direction = c("up", "down", "left", "right", "clockwise"),
  to_rotate = FALSE
)
}
\arguments{
\item{x}{scene element. Camera (object of class "scenesetr_camera") or
light (object of class "scenesetr_light") or
scene object (object of class "scenesetr_obj").}

\item{direction}{character string}

\item{to_rotate}{logical; should \code{direction} indicate the
direction of rotation were an element to be rotated about the returned axis,
or simply the direction of the returned axis?}
}
\value{
numeric vector. 3-D (x,y,z) coordinates describing a vector of unit
length in the direction specified.
}
\description{
Returns a vector that describes an axis relative to the direction and roll of
a scene element.
}
\details{
When created with default arguments, a scene element faces the positive z
direction with no roll. \code{skewer(element, "up")} therefore returns \code{c(0,1,0)},
vertically upwards. This can change as an element is rotated. \code{skewer()} is
therefore useful in determining and acting on the orientation of an element.

If \code{to_rotate} is \code{TRUE}, \code{direction} no longer describes the direction the
returned axis faces from the viewpoint of the scene element, but rather the
direction an element would rotate were it to rotate about the returned axis.
This is used in \code{\link[=rotate]{rotate()}} when \code{axis} is given as a character string.

\code{direction} can only be set to "clockwise" if \code{to_rotate} is \code{TRUE}.
}
\examples{
cam <- camera()
skewer(cam, "left")

(rolled_cam <- rotate(cam, "clockwise", 90))
(relative_left <- skewer(rolled_cam, "left"))

move(rolled_cam, relative_left)

}
