% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{spin}
\alias{spin}
\title{Spin an Object}
\usage{
spin(axis, angle, quit_after_cycle = FALSE)
}
\arguments{
\item{axis}{numeric vector or character string.
3-D (x,y,z) coordinates or one of \code{c("up", "down", "left", "right", "clockwise")}.}

\item{angle}{numeric value in degrees}

\item{quit_after_cycle}{logical value indicating if the device should be
quit once one full rotation of the element is completed}
}
\value{
A behavior function to pass to \code{\link[=behave]{behave()}}
}
\description{
Create a behavior function that rotates a scene element every frame.
}
\details{
\code{spin()} returns a behavior function for the rotation of an element each frame
about the axis and by the angle specified.
\code{axis} and \code{angle} are passed to \code{\link[=rotate]{rotate()}}.
}
