% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaves.R
\name{behaves}
\alias{behaves}
\alias{\%behaves\%}
\title{Query Behaviours of Scene Elements}
\usage{
behaves(x, behaviour)

x \%behaves\% behaviour
}
\arguments{
\item{x}{scene element or scene. Camera (object of class "scenesetr_camera")
or light (object of class "scenesetr_light") or scene object (object of class
"scenesetr_obj") or scene (object of class "scenesetr_scene").}

\item{behaviour}{character variable or variable name}
}
\value{
Boolean indicating whether the scene element has the behaviour.
}
\description{
Tests if a named behaviour is present in the behaviours of a scene element.
If a scene is supplied, \code{behaves()} will return a vector of booleans,
one for each element in the scene. \verb{\%behaves\%} and \code{behaves()} are synonyms.
}
\details{
The \code{behaviour} argument can take either a character variable of the name of
the behaviour to be queried, or a variable name to be read by \code{\link[=match.call]{match.call()}}.
This is checked for equality against the names of behaviour list of the
scene element.

The boolean vector output of \code{behaves()} when called on a scene can be used
to subset a scene.
}
\seealso{
\code{\link[=behave]{behave()}}, \code{\link[=spin]{spin()}}.
}
