% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{The Color of a Scene Element}
\usage{
paint(x, colors, scale = 0, ...)
}
\arguments{
\item{x}{scene object (object of class "scenesetr_obj") or light (object of
class "scenesetr_light") or scene (object of class "scenesetr_scene")}

\item{colors}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}

\item{scale}{numeric vector. How far along the color gradient is each face?}

\item{...}{arguments to pass to \code{colorRamp}.}
}
\value{
Scene object or light or scene with updated color.
}
\description{
Change the color of a light or scene object.
}
\details{
Paints all polygons of a scene object a specified color, or alternatively
sets the color of a light. Transparency is ignored for lights.

If more than one color is supplied to \code{colors} when painting a scene object,
a \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}} is created and called on the \code{scale},
a numeric vector with an element for each polygon.

By default, \code{\link[=light]{light()}} creates a white light, and \code{\link[=read_obj]{read_obj()}} and \code{\link[=st_as_obj]{st_as_obj()}}
create white scene objects.

If \code{x} is a scene, \code{paint()} will be applied to every element in the scene.
}
