% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Interactive 3-D Render of a Scene}
\usage{
record(
  x,
  width = 1920,
  height = 1080,
  save_to_png = FALSE,
  filename = "Rplot\%05d.png"
)
}
\arguments{
\item{x}{scene (object of class "scenesetr_scene")
or recording (object of class "scenesetr_recording")}

\item{width}{numeric width of window in pixels}

\item{height}{numeric height of window in pixels}

\item{save_to_png}{logical value. Should every frame be saved as a PNG file?}

\item{filename}{the path of the output PNG file. The frame number is substituted
if a C integer format is included in the character string.}
}
\value{
Object of class "scenesetr_recording", invisibly. List of three elements:
\itemize{
\item \code{initial_scene}: the original scene passed to \code{record()},
\item \code{final_scene}: the scene as it was in the last frame before quitting the device,
\item \code{inputs}: a list of key inputs, with one element per frame recorded.
}
}
\description{
View a scene from the perspective of a camera. Record and replay how behaviours
play out frame by frame with real-time key inputs from the user.
}
\details{
Creates a window displaying a render for each frame. Use Esc to close the window.

Supplying a scene with no camera is an error. A scene with no lights
or no scene objects will appear blank.
}
\seealso{
\code{\link[=scene]{scene()}}, \code{\link[=read_obj]{read_obj()}}, \code{\link[=record_gif]{record_gif()}}.
}
