% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_gif.R
\name{record_gif}
\alias{record_gif}
\title{Record 3-D Render to GIF}
\usage{
record_gif(
  x,
  gif_file = "animation.gif",
  width = 800,
  height = 600,
  delay = 1/30,
  loop = TRUE,
  progress = TRUE
)
}
\arguments{
\item{x}{scene (object of class "scenesetr_scene")
or recording (object of class "scenesetr_recording")}

\item{gif_file}{output gif file}

\item{width}{gif width in pixels}

\item{height}{gif height in pixel}

\item{delay}{time to show each image in seconds}

\item{loop}{if the gif should be repeated. Set to FALSE to only play
once, or a number to indicate how many times to repeat after the first.}

\item{progress}{print some verbose status output}
}
\value{
Object of class "scenesetr_recording", invisibly. List of three elements:
\itemize{
\item \code{initial_scene}: the original scene passed to \code{record()},
\item \code{final_scene}: the scene as it was in the last frame before quitting the device,
\item \code{key_inputs}: a list of key inputs, with one element per frame recorded.
}
}
\description{
Save a recording to GIF or record a scene to GIF.
}
\details{
Creates a temporary directory to store PNG files of each frame, which are
then converted to GIF by \code{\link[gifski:gifski]{gifski::gifski()}}. Frames are run using \code{\link[=record]{record()}}
with the \code{\link[grDevices:png]{grDevices::png()}} device.

It is recommended that \code{x} be a recording from \code{record()} rather than a scene,
as no visible graphics window is used.
If \code{x} is a scene, key inputs are logged and used in \code{record()} as each
frame is calculated.

All arguments except for \code{x}, \code{render_order} and \code{...} are passed to \code{gifski()}, along with
the height calculated by \code{record()} from \code{width} and the aspect of the camera.
}
