% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenesetr-package.R
\docType{package}
\name{scenesetr-package}
\alias{scenesetr}
\alias{scenesetr-package}
\title{scenesetr: Construct and Render Interactive 3-Dimensional Scenes}
\description{
Interactively explore and animate custom 3-D scenes. Define behaviours of lights, cameras and objects in response to key inputs and each other, and record scenes as they pan out in real time. For a complete list of functions, use `help(package = "scenesetr")`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/scenesetr}
  \item Report bugs at \url{https://github.com/hrryt/scenesetr/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

Authors:
\itemize{
  \item Oliver Carter-Bones
}

}
\keyword{internal}
