% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenesetr-package.R
\docType{package}
\name{scenesetr-package}
\alias{scenesetr}
\alias{scenesetr-package}
\title{scenesetr: Construct and Render Interactive 3-Dimensional Scenes}
\description{
Construct 3-dimensional scenes with elements of three types: lights, cameras and scene objects. Render scenes from the perspective of a camera with real-time key inputs. Add custom element behaviours to define location, rotation, colour and more in response to key inputs and other elements. Record scenes as they pan out in an interactive window then replay to a separate device. Add scene objects from "stars" rasters or .obj files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/scenesetr}
  \item Report bugs at \url{https://github.com/hrryt/scenesetr/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

Other contributors:
\itemize{
  \item Oliver Carter-Bones [contributor]
}

}
\keyword{internal}
