% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Interactive 3-D Render of a Scene}
\usage{
see(scene, device = grDevices::X11, width = 7, ...)
}
\arguments{
\item{scene}{scene (object of class "scenesetr_scene")}

\item{device}{function. Graphics device to be used.}

\item{width}{passed to device function}

\item{...}{passed to device function}
}
\value{
Object of class "scenesetr_recording", invisibly. List of three elements:
\itemize{
\item \code{initial_scene}: the original scene passed to \code{see()},
\item \code{final_scene}: the scene as it was in the last frame before quitting the device,
\item \code{key_inputs}: a list of key inputs, with one element per frame recorded.
}
}
\description{
View a scene from the perspective of a camera. Watch behaviours play out
frame by frame with real-time key inputs from the user.
}
\details{
Only available on windows due to the method of obtaining key inputs. This
may change in the future.

A scene must have a camera for \code{see()} to function. A scene with no lights
or no scene objects will appear blank.

Although key inputs can be used in the behaviours of scene elements, "R" and
"Q" are reserved for resetting the scene and quitting the device respectively.
}
\seealso{
\code{\link[=scene]{scene()}}, \code{\link[=read_obj]{read_obj()}}, \code{\link[=replay]{replay()}}.
}
