% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_examples.R
\name{spin}
\alias{spin}
\alias{quit_after_frame}
\alias{boid}
\alias{excited}
\alias{pushable}
\alias{bullet}
\alias{fragile}
\alias{wasd}
\alias{arrow_keys}
\alias{follow}
\alias{float_behind}
\alias{jump}
\alias{tilt}
\title{Example Behaviours of Scene Elements}
\usage{
spin(axis, angle, quit_after_cycle = FALSE)

quit_after_frame(element, frame, ...)

boid(element, scene, ...)

excited(element, scene, ...)

pushable(element, scene, ...)

bullet(element, scene, keys, ...)

fragile(element, scene, ...)

wasd(element, keys, ...)

arrow_keys(element, keys, ...)

follow(element, scene, ...)

float_behind(element, scene, ...)

jump(element, keys, ...)

tilt(element, keys, last_keys, ...)
}
\arguments{
\item{axis}{the axis or direction of rotation to be passed to the \code{axis}
argument of \code{rotate()}}

\item{angle}{the angle of rotation each frame to be passed to the \code{angle}
argument of \code{rotate()}}

\item{quit_after_cycle}{logical value indicating if the device should be
quit once one full rotation of the element is completed}

\item{element}{a scene element with the defined behaviour, to be modified by
the behaviour each frame}

\item{frame}{numeric value. Frame number, starting at one.}

\item{...}{ignored}

\item{scene}{the entire scene as it was the previous frame}

\item{keys}{character vector of keys held during the current frame}

\item{last_keys}{character vector of keys held during the previous frame}
}
\value{
For \code{spin()}, a behaviour function.

For every other function, the modified scene element, to be modified
by further behaviours or rendered in a given frame.
}
\description{
Functions to be provided to \code{\link[=behave]{behave()}} in defining the behaviours of a
scene element, or functions that return behaviour functions when called.
}
\details{
It is encouraged to peruse the code bodies of the behaviour functions
in this topic to learn how to define your own behaviours.

All of the functions in this topic except for \code{spin()} are behaviours.

\code{spin()} returns a behaviour function for the rotation of an element each frame
about the axis and by the angle specified.

An element with the \code{quit_after_frame} behaviour quits the device after one frame.

A swarm of elements with the \code{boid} behaviour will act as a swarm of boids.

An element with the \code{excited} behaviour will wiggle vertically if a player
(any element with the \code{wasd} behaviour) is not looking at the object.

An element with the \code{pushable} behaviour can be pushed horizontally by a player.

An element with the \code{bullet} behaviour will 'break' any element with the \code{fragile}
behaviour once in range, upon which the bullet element will be unplaced.
The element can be fired from a player using the 'e' key. 'breaking' an
element removes all its behaviours except for those named \code{pushable}, and
points the element upwards.

An element with the \code{wasd} behaviour can be moved forward and backward with
the 'w' and 's' keys, and can be rotated left and right with the 'a' and 'd'
keys. The \code{arrow_keys} behaviours allows the same control except using the arrow
keys.

An element with the \code{follow} behaviour follows behind and above a player, with
the same orientation as the player except for a thirty degree negative pitch.

An element with the \code{float_behind} behaviour floats behind and above a player
with a delayed effect and the same orientation as the player.

An element with the \code{jump} behaviour can jump vertically from a horizontal
floor defined by the initial position of the object when the ' ' key is held.

An element with the \code{tilt} behaviour can tilt down and up
on the press of keys '1' and '2' respectively.
}
\seealso{
\code{\link[=behave]{behave()}}, \code{\link[=behaves]{behaves()}}.
}
