% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_obj.R
\name{st_as_obj}
\alias{st_as_obj}
\title{Convert Raster to Scene Object}
\usage{
st_as_obj(x)
}
\arguments{
\item{x}{stars raster object}
}
\value{
Scene object (object of class "scenesetr_obj").
}
\description{
Convert a stars raster object to a scene object with xz coordinates defined
by the raster grid and y coordinates defined by a raster attribute.
}
\details{
The raster object is assumed to have 2 dimensions and 1 attribute.
The x and y dimensions of the raster are mapped onto the x and z axes in 3-D space.
The attribute of the raster is taken to be the y coordinate at each cell in the xz grid.

The raster object is first converted by \link[stars:st_as_sf]{stars::st_as_sf} to an sf object
of square polygons, one for each raster cell, with one attribute value per polygon.
The y coordinate of each point is determined by the attribute value of the
polygon in which it is first encountered.

Surface normals are added to the scene object by \link{add_normals}.

The returned scene object is painted white, unplaced, has no behaviours and
faces the positive z direction.
}
\seealso{
\code{\link[=read_obj]{read_obj()}}, \code{\link[=scene]{scene()}}, \code{\link[=see]{see()}}.
}
